+++ data/src/file_fs/__init__.py
@@ -25,6 +25,9 @@ __copyright__ = "Copyright (c) 2008-2024 Hive Solutions Lda."
 __license__ = "Apache License, Version 2.0"
 """ The license for the module """
 
+from . import mocks
 from . import system
+from . import test
 
 from .system import FileFS
+from .test import FileFSTest