+++ data/src/file_fs/mocks.py
@@ -19,15 +19,21 @@
 # You should have received a copy of the Apache License along with
 # Hive Colony Framework. If not, see <http://www.apache.org/licenses/>.
 
+__author__ = "João Magalhães <joamag@hive.pt>"
+""" The author(s) of the module """
+
 __copyright__ = "Copyright (c) 2008-2024 Hive Solutions Lda."
 """ The copyright for the module """
 
 __license__ = "Apache License, Version 2.0"
 """ The license for the module """
 
-from . import mocks
-from . import system
-from . import test
 
-from .system import FileZip
-from .test import FileZipTest
+class MockConnection(object):
+    """
+    Mock connection, holding the (engine) file connection
+    that is used by the file engine operations.
+    """
+
+    def __init__(self, file_connection):
+        self.file_connection = file_connection