+++ data/src/file_gridfs_plugin.py
@@ -90,3 +90,9 @@ class FileGridFSPlugin(colony.Plugin):
 
     def mtime(self, connection, file_name):
         return self.system.mtime(connection, file_name)
+
+    def exists(self, connection, file_name):
+        return self.system.exists(connection, file_name)
+
+    def exists_directory(self, connection, directory_name):
+        return self.system.exists_directory(connection, directory_name)